fix(pvs): resolve Low Level (Level 3) static analyzer warnings and formatting notes
- Explicitly cast int/char arguments to match fprintf format specifiers in pangen2.c and pangen3.c
- Add NULL pointer checks in claim_has_accept() in pangen7.c
- Guard pointer dereferences in getlocal() and setlocal() in sched.c
- Guard NULL pointers when building struct nodes in structs.c
- Guard NULL pointers in dupnode() and isequal() in tl_cache.c
- Cast memory size arguments to long for printf formatting in tl_cache.c and tl_main.c

+++ Src/pangen2.c
@@ -2219,7 +2219,7 @@ scan_seq(Sequence *s)
 			&& !(f->status & L_ATOM)
 			&& !(g->status & (ATOM|L_ATOM)))
 #endif
-			{	fprintf(fd_tt, "\t/* mark-down line %d status %d = %d */\n", f->n->ln, f->status, (f->status & D_ATOM));
+			{	fprintf(fd_tt, "\t/* mark-down line %d status %d = %d */\n", f->n->ln, f->status, (int)(f->status & D_ATOM));
 				return 1; /* assume worst case */
 		}	}
 		for (h = f->sub; h; h = h->nxt)
